% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendateseq.R
\name{genDateSeq}
\alias{genDateSeq}
\title{genDateSeq}
\usage{
genDateSeq(from, to, by = "week")
}
\arguments{
\item{from}{starting date}

\item{to}{end date}

\item{by}{increment of the sequence, default = week, options: day, week, month}
}
\value{
a vector of class "Date", in formate m/d/yyyy
}
\description{
\code{genDateSeq} geneates a list of dates for query
}
\examples{
genDateSeq("2010/1/1", "2018/1/1", "year")
genDateSeq("2018/1/1", "2018/3/1")
}
