% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genurl.R
\name{genUrl}
\alias{genUrl}
\title{genUrl}
\usage{
genUrl(originCity = "LAX", destinationCity = "JFK", departureDate,
  returnDate)
}
\arguments{
\item{originCity}{fly from}

\item{destinationCity}{fly to}

\item{departureDate}{date to depart in m/d/yyyy}

\item{returnDate}{date to return in m/d/yyyy}
}
\value{
query URL :: string
}
\description{
\code{genUrl} generate query url specific to the dates and air route specified by the parameters
}
\examples{
remDr <- setup()
remDr$open()
remDr$navigate(genUrl("JFK", "SEA", "1/1/2019", "2/1/2019"))
}
