% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{main}
\usage{
main(originCity, destinationCity, dateGrid, remDr, sleep = 5L)
}
\arguments{
\item{originCity}{fly from}

\item{destinationCity}{fly yo}

\item{dateGrid}{a list of departure and return dates}

\item{remDr}{remote driver selenium object}

\item{sleep}{wait time for dynamic contents, default = 5}
}
\value{
departure date | return date | price  :: dataframe
}
\description{
\code{main} scrape the site with the given data
}
\examples{
remDr <- setup()
remDr$open()
departD <- seq(as.Date("2019/1/1"), as.Date("2019/2/1"), "week")
departD <- format(departD, "\%m/\%d/20\%y")
returnD <- seq(as.Date("2019/1/2"), as.Date("2019/2/2"), "week")
returnD <- format(returnD, "\%m/\%d/20\%y")
dateGrid <- data.frame("depart"=departD, "return"=returnD)
main("LAX", "JFK", dateGrid, remDr)
}
