% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricedf.R
\docType{data}
\name{pricedf}
\alias{pricedf}
\title{Delta Airline ticket price}
\format{A data frame with 192 rows and 3 variables:
\describe{
  \item{\code{departure}}{integer. departure date}
  \item{\code{return}}{integer. return date}
  \item{\code{price}}{double. price listings}
}}
\usage{
pricedf
}
\description{
Delta Airline ticket price from LAX to JFK
}
\keyword{datasets}
